core logic completed for all stories. Added swagger api for documentation

+++ Backend/src/main/java/com/example/library_management_system/ServiceImpl/BookServiceImpl.java
@@ -7,9 +7,8 @@ import com.example.library_management_system.Service.BookService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.awt.print.Book;
+
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class BookServiceImpl implements BookService {
@@ -31,7 +30,7 @@ public class BookServiceImpl implements BookService {
     @Override
     public Books getBookbyId(int bookId) {
 
-        return bookRepository.findById(bookId).orElseThrow(() -> new RuntimeException("No Book with given book id"));
+        return bookRepository.findById((Integer)bookId).orElseThrow(() -> new RuntimeException("No Book with given book id"));
     }
 
     @Override
